SQL schema migration + YearMonthDateType util + QueryDSL

+++ src/main/java/bartkisiel/github/io/HotelBooking/room/RoomPredicates.java
@@ -10,7 +10,7 @@ public final class RoomPredicates {
 
     }
 
-    public static Predicate availableRoom(int hotelId) {
+    public static Predicate availableRoom(long hotelId) {
         return room.hotel.id.eq(hotelId).and(room.reservation.isNull());
     }
 }
